Implemented hashtag detection and redirection
